+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add TUI switch for status and priority
 - Add default due config entry
+- Support numbered priority config
 
 ### Changed